In progress
in progress

+++ backend/resourcify/src/main/java/com/resourcifyproject/resourcify/ResourcifyUser.java
@@ -1,13 +1,10 @@
 package com.resourcifyproject.resourcify;
 
-import java.util.Arrays;
-import java.util.List;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.stream.Collectors;
 
 import static java.lang.String.valueOf;
 
@@ -19,7 +16,7 @@ public class ResourcifyUser implements UserDetails {
     private final String password;
     private final String firstname;
     private final String lastname;
-    private final int availablefunds;
+    private final float availablefunds;
     private Collection<SimpleGrantedAuthority> authorities = new ArrayList<SimpleGrantedAuthority>();
 
     public ResourcifyUser(User user) {
